fix(auth): review fixes and align user model with Prisma
