--- backend/src/auth/decorators/roles.decorator.ts
@@ -1,5 +1,0 @@
-import { SetMetadata } from '@nestjs/common';
-import { UserRole } from '../types/auth.types';
-
-export const ROLES_KEY = 'roles';
-export const Roles = (...roles: UserRole[]) => SetMetadata(ROLES_KEY, roles);
