+++ backend/src/auth/guards/director.guard.ts
@@ -1,0 +1,22 @@
+import {
+  CanActivate,
+  ExecutionContext,
+  ForbiddenException,
+  Injectable,
+} from '@nestjs/common';
+import { AuthenticatedUser } from '../types/auth.types';
+
+@Injectable()
+export class DirectorGuard implements CanActivate {
+  canActivate(context: ExecutionContext): boolean {
+    const { user } = context
+      .switchToHttp()
+      .getRequest<{ user?: AuthenticatedUser }>();
+
+    if (!user?.isDirector) {
+      throw new ForbiddenException();
+    }
+
+    return true;
+  }
+}
