--- backend/src/auth/guards/roles.guard.ts
@@ -1,35 +1,0 @@
-import {
-  CanActivate,
-  ExecutionContext,
-  ForbiddenException,
-  Injectable,
-} from '@nestjs/common';
-import { Reflector } from '@nestjs/core';
-import { ROLES_KEY } from '../decorators/roles.decorator';
-import { AuthenticatedUser, UserRole } from '../types/auth.types';
-
-@Injectable()
-export class RolesGuard implements CanActivate {
-  constructor(private readonly reflector: Reflector) {}
-
-  canActivate(context: ExecutionContext): boolean {
-    const requiredRoles = this.reflector.getAllAndOverride<UserRole[]>(
-      ROLES_KEY,
-      [context.getHandler(), context.getClass()],
-    );
-
-    if (!requiredRoles || requiredRoles.length === 0) {
-      return true;
-    }
-
-    const { user } = context
-      .switchToHttp()
-      .getRequest<{ user?: AuthenticatedUser }>();
-
-    if (!user || !requiredRoles.includes(user.role)) {
-      throw new ForbiddenException();
-    }
-
-    return true;
-  }
-}
