+++ docs/auth.md
@@ -76,7 +76,7 @@ POST /api/v1/auth/refresh   (no body)             →  { accessToken, user }
 GET  /api/v1/auth/me        (Bearer token)        →  user
 ```
 
-`user` contains: `id`, `email`, `firstName`, `lastName`, `role`, `accountStatus`, `mustChangePassword`.
+`user` contains: `id`, `email`, `firstName`, `lastName`, `isDirector`, `accountStatus`, `mustChangePassword`.
 
 ## Things that will trip you up
 
@@ -84,5 +84,6 @@ GET  /api/v1/auth/me        (Bearer token)        →  user
 - **The refresh cookie is scoped to `/api/v1/auth/refresh`.** The browser will not send it to any other path. That is deliberate.
 - **CORS:** the API only accepts requests from the origin set in `CORS_ORIGIN` (`http://localhost:5173` in development). A different port means the request is rejected before it reaches any of this.
 - **`mustChangePassword: true`** means the user is still on an admin-issued password and must set their own before doing anything else.
+- **`isDirector` is not a role.** It is a global admin flag, and it is the only permission this endpoint returns. Project roles (coordinator, executor, partner) are per-project and will arrive on a separate endpoint once projects exist — do not expect them here.
 - **`sameSite: 'strict'` and production domains:** the refresh cookie is only sent when the request comes from the same site. This is fine in development (`localhost:5173` → `localhost:3000`). If the frontend and the API ever end up on genuinely different domains, the browser will silently stop sending the cookie and session refresh will break with no error anywhere. Check this before the first deployment.
